Don't store state every block, when syncing or rescanning
A major bottle neck during the initial synchronization, rescanning or processing new blocks is due to storing the consensus state to disk.

This is done, so the state can easily be rolled back in case of a block reorganization.

However, during the inital synchronization or when rescanning and reparsing historical blocks, it is only needed in case of a shutdown or crash, but can be skipped otherwise.

Also: store state every 5000 blocks during initial synchronization or rescanning.